+++ innsender/src/test/kotlin/no/nav/soknad/innsending/utils/builders/tilleggsstonad/JsonDagligReiseTestBuilder.kt
@@ -17,33 +17,33 @@ class JsonDagligReiseTestBuilder {
 	private var sluttdatoDdMmAaaa: String = Date.formatToLocalDate(LocalDateTime.now().plusMonths(1))
 	private var soknadsPeriode: SoknadsPeriode =
 		SoknadsPeriode(startdato = startdatoDdMmAaaa, sluttdato = sluttdatoDdMmAaaa)
-	private var hvorMangeReisedagerHarDuPerUke: Int = 5
+	private var hvorMangeReisedagerHarDuPerUke: Double = 5.0
 	private var harDuEnReiseveiPaSeksKilometerEllerMer: String = "Ja"
 	private var harDuAvMedisinskeArsakerBehovForTransportUavhengigAvReisensLengde: String? = null
-	private var hvorLangReiseveiHarDu: Int = 10
+	private var hvorLangReiseveiHarDu: Double = 10.0
 	private var velgLand1: VelgLand = VelgLand(
 		label = "Norge", value = "NO"
 	)
 	private var adresse1: String = "Kongensgate 10"
 	private var postnr1: String? = "3701"
 	private var kanDuReiseKollektivtDagligReise: String = "Nei"
-	private var hvilkeUtgifterHarDuIforbindelseMedReisenDagligReise: Int? = null
+	private var hvilkeUtgifterHarDuIforbindelseMedReisenDagligReise: Double? = null
 	private var hvaErHovedarsakenTilAtDuIkkeKanReiseKollektivt: String? = "hentingEllerLeveringAvBarn"
 	private var beskrivDeSpesielleForholdeneVedReiseveienSomGjorAtDuIkkeKanReiseKollektivt: String? =
 		"Mange transportbytter, ekstra lang reisetid"
 	private var hvilkeAndreArsakerErDetSomGjorAtDuIkkeKanReiseKollektivt: String? = "Ingen"
 	private var kanBenytteEgenBil: KanBenytteEgenBil? = KanBenytteEgenBil(
-		bompenger = 150,
-		piggdekkavgift = 1000,
+		bompenger = 150.0,
+		piggdekkavgift = 1000.0,
 		ferje = null,
 		annet = null,
 		vilDuHaUtgifterTilParkeringPaAktivitetsstedet = "JA",
-		parkering = 200,
+		parkering = 200.0,
 		hvorOfteOnskerDuASendeInnKjoreliste = "UKE"
 	)
 	private var kanIkkeBenytteEgenBil: KanIkkeBenytteEgenBil? = null
 	private var kanDuBenytteDrosje: String? = null
-	private var oppgiDenTotaleKostnadenDuHarTilBrukAvDrosjeIPeriodenDuSokerOmStonadFor: Int? = null
+	private var oppgiDenTotaleKostnadenDuHarTilBrukAvDrosjeIPeriodenDuSokerOmStonadFor: Double? = null
 	private var hvorforKanDuIkkeBenytteDrosje: String? = null
 	private var kanIkkeReiseKollektivtDagligReise: KanIkkeReiseKollektivt? =
 		buildKanIkkeReiseKollektivtDagligReise()
@@ -52,7 +52,7 @@ class JsonDagligReiseTestBuilder {
 		soknadsPeriode = SoknadsPeriode(startdato, sluttdato)
 	}
 
-	fun hvorMangeReisedagerHarDuPerUke(hvorMangeReisedagerHarDuPerUke: Int) =
+	fun hvorMangeReisedagerHarDuPerUke(hvorMangeReisedagerHarDuPerUke: Double) =
 		apply { this.hvorMangeReisedagerHarDuPerUke = hvorMangeReisedagerHarDuPerUke }
 
 	fun harDuEnReiseveiPaSeksKilometerEllerMer(harDuEnReiseveiPaSeksKilometerEllerMer: String) =
@@ -68,14 +68,14 @@ class JsonDagligReiseTestBuilder {
 	fun hvaErHovedarsakenTilAtDuIkkeKanReiseKollektivt(hvaErHovedarsakenTilAtDuIkkeKanReiseKollektivt: String) =
 		apply { this.hvaErHovedarsakenTilAtDuIkkeKanReiseKollektivt = hvaErHovedarsakenTilAtDuIkkeKanReiseKollektivt }
 
-	fun hvorLangReiseveiHarDu(hvorLangReiseveiHarDu: Int) = apply { this.hvorLangReiseveiHarDu = hvorLangReiseveiHarDu }
+	fun hvorLangReiseveiHarDu(hvorLangReiseveiHarDu: Double) = apply { this.hvorLangReiseveiHarDu = hvorLangReiseveiHarDu }
 	fun velgLand1(velgLand1: VelgLand) = apply { this.velgLand1 = velgLand1 }
 	fun adresse1(adresse1: String) = apply { this.adresse1 = adresse1 }
 	fun postnr1(postnr1: String?) = apply { this.postnr1 = postnr1 }
 	fun kanDuReiseKollektivtDagligReise(kanDuReiseKollektivtDagligReise: String) =
 		apply { this.kanDuReiseKollektivtDagligReise = kanDuReiseKollektivtDagligReise }
 
-	fun hvilkeUtgifterHarDuIforbindelseMedReisenDagligReise(hvilkeUtgifterHarDuIforbindelseMedReisenDagligReise: Int?) =
+	fun hvilkeUtgifterHarDuIforbindelseMedReisenDagligReise(hvilkeUtgifterHarDuIforbindelseMedReisenDagligReise: Double?) =
 		apply {
 			this.hvilkeUtgifterHarDuIforbindelseMedReisenDagligReise = hvilkeUtgifterHarDuIforbindelseMedReisenDagligReise
 		}
@@ -89,7 +89,7 @@ class JsonDagligReiseTestBuilder {
 
 	fun kanDuBenytteDrosje(kanDuBenytteDrosje: String?) = apply { this.kanDuBenytteDrosje = kanDuBenytteDrosje }
 	fun oppgiDenTotaleKostnadenDuHarTilBrukAvDrosjeIPeriodenDuSokerOmStonadFor(
-		oppgiDenTotaleKostnadenDuHarTilBrukAvDrosjeIPeriodenDuSokerOmStonadFor: Int?
+		oppgiDenTotaleKostnadenDuHarTilBrukAvDrosjeIPeriodenDuSokerOmStonadFor: Double?
 	) = apply {
 		this.oppgiDenTotaleKostnadenDuHarTilBrukAvDrosjeIPeriodenDuSokerOmStonadFor =
 			oppgiDenTotaleKostnadenDuHarTilBrukAvDrosjeIPeriodenDuSokerOmStonadFor
